Split into 2 classes
